убрала лишние импорты

+++ src/main/java/com/example/contactlistjdbc/listener/InitContact.java
@@ -1,13 +1,10 @@
 package com.example.contactlistjdbc.listener;
 
-import com.example.contactlistjdbc.controller.ContactController;
 import com.example.contactlistjdbc.data.Contact;
 import com.example.contactlistjdbc.repository.ContactRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.boot.context.event.ApplicationStartedEvent;
-import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
@@ -29,7 +26,7 @@ public class InitContact {
     for ( int i = 0; i < 10; i++ ) {
       int value = i + 1;
       Contact contact = new Contact();
-      contact.setId( (long) value );
+      contact.setId( value );
       contact.setFirstName( "FirstName " + i );
       contact.setLastName( "LastName " + i );
       contact.setEmail( "sdf" + i + "@mail.ru" );
